+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ### Bug Fixes
 
 - remove unused vars to pass ESLint ([371bba7](https://github.com/involvex/awesome-github-app/commit/371bba75e4f95aeac7e4f0e25f2193b7b6885caa))
+- rename API route for Expo Router compatibility ([e921945](https://github.com/involvex/awesome-github-app/commit/e9219457486bf32f4eef4b14c7f9c305b75a123d))
 
 ### Features
 
 - add OAuth config and improve type safety in explore screens ([0e8ae38](https://github.com/involvex/awesome-github-app/commit/0e8ae38559c0b353c2a37a4577f84b337378172e))
 - add useActivity hook, Plan.md, and update gitignore ([b4acdc6](https://github.com/involvex/awesome-github-app/commit/b4acdc65da1a0e6dde7c17f0c28cea6727b167fb))
+- add web OAuth credentials to app.json ([82179f7](https://github.com/involvex/awesome-github-app/commit/82179f7715786b71554aa73f0b8655d42d12d101))
 - enhance home screen UI and replace SecureStore with custom storage ([0437282](https://github.com/involvex/awesome-github-app/commit/0437282d5a150a390a6a87762534136d80654107))
 - integrate Expo splash screen and add adaptive Android app icons ([b2fedb7](https://github.com/involvex/awesome-github-app/commit/b2fedb7ae362f4368cdde7d8c920355d8f4a10ec))
+- **oauth:** add Cloudflare Worker for secure web OAuth token exchange ([3b87a9e](https://github.com/involvex/awesome-github-app/commit/3b87a9e5c0426c25c381524d7e364d55b3e66e93))